requirements.txt
discord.py
python-dotenv
openai
yt-dlp
flask
aiofiles
PyNaCl
ffmpeg-python